backend/requirements.txt
fastapi
uvicorn
pillow
tweepy
